requirements.txt
aniso8601==8.0.0
attrs==19.3.0
certifi==2019.11.28
Click==7.0
colorama==0.4.3
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Neo4j4==4.5.1
Flask-RESTful==0.3.7
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-WTF==0.14.2
future==0.18.2
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
MarkupSafe==1.1.1
neobolt==1.7.16
neotime==1.7.4
prompt-toolkit==2.0.10
py2neo==4.3.0
Pygments==2.3.1
pyrsistent==0.15.7
pytz==2019.3
six==1.14.0
urllib3==1.24.3
visitor==0.1.3
wcwidth==0.1.8
Werkzeug==0.16.0
WTForms==2.2.1
zipp==2.1.0
